adding create order endpoint

+++ src/main/java/com/mini_project/miniproject/events/dto/CreateEventVoucherDto.java
@@ -2,10 +2,12 @@ package com.mini_project.miniproject.events.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class CreateEventVoucherDto {
     private String code;
-    private Double discountPercentage;
+    private BigDecimal discountPercentage;
     private String startDate;
     private String endDate;
 }
